+++ py-polars/polars/io/ipc/functions.py
@@ -10,7 +10,7 @@ from polars.io._utils import _prepare_file_arg
 from polars.utils.various import normalise_filepath
 
 with contextlib.suppress(ImportError):
-    from polars.polars import ipc_schema as _ipc_schema
+    from polars.polars import read_ipc_schema as _read_ipc_schema
 
 if TYPE_CHECKING:
     from io import BytesIO
@@ -128,7 +128,7 @@ def read_ipc_schema(source: str | BinaryIO | Path | bytes) -> dict[str, PolarsDa
     if isinstance(source, (str, Path)):
         source = normalise_filepath(source)
 
-    return _ipc_schema(source)
+    return _read_ipc_schema(source)
 
 
 def scan_ipc(
