+++ py-polars/src/functions/mod.rs
@@ -1,3 +1,4 @@
 pub mod eager;
+pub mod io;
 pub mod lazy;
 pub mod whenthen;
