+++ py-polars/src/lib.rs
@@ -47,7 +47,6 @@ use polars_core::POOL;
 use pyo3::exceptions::PyValueError;
 use pyo3::panic::PanicException;
 use pyo3::prelude::*;
-use pyo3::types::PyDict;
 use pyo3::wrap_pyfunction;
 
 use crate::conversion::Wrap;
@@ -57,7 +56,6 @@ use crate::error::{
     NoDataError, PyPolarsErr, SchemaError, SchemaFieldNotFoundError, StructFieldNotFoundError,
 };
 use crate::expr::PyExpr;
-use crate::file::{get_either_file, EitherRustPythonFile};
 use crate::lazyframe::PyLazyFrame;
 use crate::lazygroupby::PyLazyGroupBy;
 use crate::prelude::DataType;
@@ -91,44 +89,6 @@ fn enable_string_cache(toggle: bool) {
 #[pyfunction]
 fn using_string_cache() -> bool {
     polars_rs::using_string_cache()
-}
-
-#[cfg(feature = "ipc")]
-#[pyfunction]
-fn ipc_schema(py: Python, py_f: PyObject) -> PyResult<PyObject> {
-    use polars_core::export::arrow::io::ipc::read::read_file_metadata;
-    let metadata = match get_either_file(py_f, false)? {
-        EitherRustPythonFile::Rust(mut r) => {
-            read_file_metadata(&mut r).map_err(PyPolarsErr::from)?
-        }
-        EitherRustPythonFile::Py(mut r) => read_file_metadata(&mut r).map_err(PyPolarsErr::from)?,
-    };
-
-    let dict = PyDict::new(py);
-    for field in metadata.schema.fields {
-        let dt: Wrap<DataType> = Wrap((&field.data_type).into());
-        dict.set_item(field.name, dt.to_object(py))?;
-    }
-    Ok(dict.to_object(py))
-}
-
-#[cfg(feature = "parquet")]
-#[pyfunction]
-fn parquet_schema(py: Python, py_f: PyObject) -> PyResult<PyObject> {
-    use polars_core::export::arrow::io::parquet::read::{infer_schema, read_metadata};
-
-    let metadata = match get_either_file(py_f, false)? {
-        EitherRustPythonFile::Rust(mut r) => read_metadata(&mut r).map_err(PyPolarsErr::from)?,
-        EitherRustPythonFile::Py(mut r) => read_metadata(&mut r).map_err(PyPolarsErr::from)?,
-    };
-    let arrow_schema = infer_schema(&metadata).map_err(PyPolarsErr::from)?;
-
-    let dict = PyDict::new(py);
-    for field in arrow_schema.fields {
-        let dt: Wrap<DataType> = Wrap((&field.data_type).into());
-        dict.set_item(field.name, dt.to_object(py))?;
-    }
-    Ok(dict.to_object(py))
 }
 
 #[pyfunction]
@@ -260,6 +220,30 @@ fn polars(py: Python, m: &PyModule) -> PyResult<()> {
     m.add_wrapped(wrap_pyfunction!(functions::whenthen::when))
         .unwrap();
 
+    // Functions - I/O
+    #[cfg(feature = "ipc")]
+    m.add_wrapped(wrap_pyfunction!(functions::io::read_ipc_schema))
+        .unwrap();
+    #[cfg(feature = "parquet")]
+    m.add_wrapped(wrap_pyfunction!(functions::io::read_parquet_schema))
+        .unwrap();
+
+    // Functions - utility
+    m.add_wrapped(wrap_pyfunction!(get_polars_version)).unwrap();
+    m.add_wrapped(wrap_pyfunction!(get_index_type)).unwrap();
+    m.add_wrapped(wrap_pyfunction!(threadpool_size)).unwrap();
+
+    // Functions - other
+    m.add_wrapped(wrap_pyfunction!(dtype_str_repr)).unwrap();
+    m.add_wrapped(wrap_pyfunction!(enable_string_cache))
+        .unwrap();
+    m.add_wrapped(wrap_pyfunction!(using_string_cache)).unwrap();
+    m.add_wrapped(wrap_pyfunction!(set_float_fmt)).unwrap();
+    m.add_wrapped(wrap_pyfunction!(get_float_fmt)).unwrap();
+    #[cfg(feature = "object")]
+    m.add_wrapped(wrap_pyfunction!(register_object_builder))
+        .unwrap();
+
     // Exceptions
     m.add("ArrowError", py.get_type::<ArrowErrorException>())
         .unwrap();
@@ -291,23 +275,7 @@ fn polars(py: Python, m: &PyModule) -> PyResult<()> {
     )
     .unwrap();
 
-    // Other
-    m.add_wrapped(wrap_pyfunction!(dtype_str_repr)).unwrap();
-    m.add_wrapped(wrap_pyfunction!(get_polars_version)).unwrap();
-    m.add_wrapped(wrap_pyfunction!(enable_string_cache))
-        .unwrap();
-    m.add_wrapped(wrap_pyfunction!(using_string_cache)).unwrap();
-    #[cfg(feature = "ipc")]
-    m.add_wrapped(wrap_pyfunction!(ipc_schema)).unwrap();
-    #[cfg(feature = "parquet")]
-    m.add_wrapped(wrap_pyfunction!(parquet_schema)).unwrap();
-    m.add_wrapped(wrap_pyfunction!(threadpool_size)).unwrap();
-    m.add_wrapped(wrap_pyfunction!(get_index_type)).unwrap();
-    m.add_wrapped(wrap_pyfunction!(set_float_fmt)).unwrap();
-    m.add_wrapped(wrap_pyfunction!(get_float_fmt)).unwrap();
-    #[cfg(feature = "object")]
-    m.add_wrapped(wrap_pyfunction!(register_object_builder))
-        .unwrap();
+    // Build info
     #[cfg(feature = "build_info")]
     m.add(
         "_build_info_",
